Allow users to cycle start/end of index date. (#811)
Supports selecting the event start/end.

For backwards compatibility, it was necessary to allow 'falsy' json properties to be serialized out for the EndWindow of AdditionalCriteria.

The absence of a value (null) means we should use the event's end date.  When the value is 'true', we should also use it. it is only when the value is explicitly 'false', do we not use the end date. Therefore, we allow the value of UseEventEnd to be serialzied to JSON even if 'false'.